Extract CommentCleaner
