+++ common/app/model/dotcomrendering/pageElements/PageElement.scala
@@ -1170,42 +1170,3 @@ object PageElement {
 
   val pageElementWrites: Writes[PageElement] = Json.writes[PageElement]
 }
-
-// ------------------------------------------------------
-// Misc.
-// ------------------------------------------------------
-
-object Cleaner {
-  def split(html: String): List[(String, String)] = {
-    Jsoup
-      .parseBodyFragment(html)
-      .body()
-      .children()
-      .asScala
-      .toList
-      .map(el => (el.tagName, el.outerHtml))
-  }
-}
-
-object CommentCleaner {
-  def getBody(html: String): String = {
-    Jsoup
-      .parseBodyFragment(html)
-      .getElementsByClass("d2-body")
-      .html()
-  }
-
-  def getAvatar(html: String): String = {
-    Jsoup
-      .parseBodyFragment(html)
-      .getElementsByClass("d2-avatar")
-      .attr("src")
-  }
-
-  def getDateTime(html: String): String = {
-    Jsoup
-      .parseBodyFragment(html)
-      .getElementsByClass("d2-datetime")
-      .html()
-  }
-}
